+++ docs/examples/styles/grid_columns.py
@@ -19,4 +19,4 @@ class MyApp(App):
         )
 
 
-app = MyApp(css_path="grid_columns.css")
+app = MyApp(css_path="grid_columns.tcss")