+++ docs/examples/styles/link_color.py
@@ -22,4 +22,4 @@ class LinkColorApp(App):
         )
 
 
-app = LinkColorApp(css_path="link_color.css")
+app = LinkColorApp(css_path="link_color.tcss")