+++ CHANGELOG.md
@@ -7,7 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
-- Add `min-height: 1;` to `LoadingIndicator`'s `DEFAULT_CSS`.
+### Fixed
+
+- Could not hide/show/disable/enable tabs in nested `TabbedContent` https://github.com/Textualize/textual/pull/3150
+
+## [0.34.0] - 2023-08-22
 
 ### Added
 
@@ -22,6 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 - grid-columns and grid-rows now accept an `auto` token to detect the optimal size https://github.com/Textualize/textual/pull/3107
+- LoadingIndicator now has a minimum height of 1 line.
 
 ### Fixed
 
@@ -1197,6 +1202,7 @@ https://textual.textualize.io/blog/2022/11/08/version-040/#version-040
 - New handler system for messages that doesn't require inheritance
 - Improved traceback handling
 
+[0.34.0]: https://github.com/Textualize/textual/compare/v0.33.0...v0.34.0
 [0.33.0]: https://github.com/Textualize/textual/compare/v0.32.0...v0.33.0
 [0.32.0]: https://github.com/Textualize/textual/compare/v0.31.0...v0.32.0
 [0.31.0]: https://github.com/Textualize/textual/compare/v0.30.0...v0.31.0